+++ hydra-node/test/Hydra/Chain/Direct/TxSpec.hs
@@ -52,7 +52,8 @@ spec =
             forAll genScriptRegistry $ \scriptRegistry ->
               let params = HeadParameters cperiod allParties
                   tx = initTx testNetworkId cardanoKeys params txIn
-               in case observeInitTx testNetworkId cardanoKeys cperiod party parties tx of
+               in case first NotAnInit (observeRawInitTx testNetworkId tx)
+                    >>= (first NotAnInitForUs . observeInitTx cardanoKeys cperiod party parties) of
                     Right InitObservation{initials, threadOutput} -> do
                       let InitialThreadOutput{initialThreadUTxO = (headInput, headOutput, headDatum)} = threadOutput
                           initials' = Map.fromList [(a, (b, c)) | (a, b, c) <- initials]
@@ -104,12 +105,13 @@ spec =
               -- construct different/wrong CP
               wrongCPeriod = UnsafeContestationPeriod $ cp + wordToNatural i
               tx = initTx testNetworkId cardanoKeys params txIn
-          pure $ case observeInitTx testNetworkId cardanoKeys wrongCPeriod party parties tx of
+          pure $ case first NotAnInit (observeRawInitTx testNetworkId tx)
+            >>= (first NotAnInitForUs . observeInitTx cardanoKeys wrongCPeriod party parties) of
             Right InitObservation{} -> do
               property False
                 & counterexample "Failed to ignore init tx with the wrong contestation period."
                 & counterexample (renderTx tx)
-            Left CPMismatch -> property True
+            Left (NotAnInitForUs CPMismatch{}) -> property True
             Left e ->
               property False
                 & counterexample "Failed to ignore init tx for the right reason."
@@ -123,12 +125,13 @@ spec =
               wrongCardanoKeys = genForParty genVerificationKey <$> (wrongParty : parties)
               params = HeadParameters cperiod allParties
               tx = initTx testNetworkId cardanoKeys params txIn
-           in case observeInitTx testNetworkId wrongCardanoKeys cperiod party parties tx of
+           in case first NotAnInit (observeRawInitTx testNetworkId tx)
+                >>= (first NotAnInitForUs . observeInitTx wrongCardanoKeys cperiod party parties) of
                 Right InitObservation{} -> do
                   property False
                     & counterexample "Failed to ignore init tx with the wrong cardano keys."
                     & counterexample (renderTx tx)
-                Left PTsNotMintedCorrectly -> property True
+                Left (NotAnInitForUs PTsNotMintedCorrectly{}) -> property True
                 Left e ->
                   property False
                     & counterexample "Failed to ignore init tx for the right reason."